backend/requirements.txt
fastapi
uvicorn
groq
openai-whisper
langdetect
python-multipart
python-dotenv
